+++ Runtime/NativePriorityQueueDisposeJob.cs
@@ -1,0 +1,11 @@
+﻿using Unity.Jobs;
+
+namespace BonnFireGames.CustomNativeContainers
+{
+    internal struct NativePriorityQueueDisposeJob : IJob
+    {
+        internal NativePriorityQueueDispose Data;
+
+        public void Execute() => this.Data.Dispose();
+    }
+}